safari-534.53-branch/LayoutTests/fast/events/popup-blocked-from-untrusted-click-event-on-anchor-expected.txt

test
PASS win is undefined.
